# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/delete": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "删除用户",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "编辑用户信息",
                "parameters": [
                    {
                        "description": "编辑用户信息,适用于用户自己编辑自己的信息",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserEditReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "description": "获取用户列表",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserListRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/manager-edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "管理员编辑用户",
                "parameters": [
                    {
                        "description": "管理员编辑用户",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "管理员添加用户",
                "parameters": [
                    {
                        "description": "管理员添加用户",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.ResourceQuota": {
            "type": "object",
            "required": [
                "cpu",
                "mem"
            ],
            "properties": {
                "cpu": {
                    "description": "可以使用的cpu总核数",
                    "type": "string"
                },
                "mem": {
                    "description": "可以使用的内存总量",
                    "type": "string"
                }
            }
        },
        "model.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar_num": {
                    "type": "integer"
                },
                "create_time": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "modify_time": {
                    "type": "string"
                },
                "name": {
                    "description": "name要是唯一的，对应k8s的命名空间",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resource_quota": {
                    "$ref": "#/definitions/model.ResourceQuota"
                },
                "role_id": {
                    "description": "1表示管理员，2表示开发者",
                    "type": "integer"
                }
            }
        },
        "model.UserDeleteBase": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "is_clean_app": {
                    "description": "是否删除对应用户下的所有app",
                    "type": "boolean"
                },
                "name": {
                    "description": "要删除的用户名",
                    "type": "string"
                }
            }
        },
        "model.UserDeleteReq": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDeleteBase"
                    }
                }
            }
        },
        "model.UserEditReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "model.UserListBase": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.UserListReq": {
            "type": "object",
            "required": [
                "current",
                "pageSize"
            ],
            "properties": {
                "current": {
                    "description": "当前第几页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页多少条",
                    "type": "integer"
                },
                "role_id": {
                    "description": "角色,可不传 1表示管理员 2表示开发者",
                    "type": "integer"
                },
                "search": {
                    "description": "模糊搜索字段，可不传",
                    "type": "string"
                }
            }
        },
        "model.UserListRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.UserListBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.UserLoginReq": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
        "model.UserLoginRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.UserLoginResBase"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.UserLoginResBase": {
            "type": "object",
            "properties": {
                "avatar_num": {
                    "type": "integer"
                },
                "avatar_url": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "modify_time": {
                    "type": "string"
                },
                "name": {
                    "description": "name要是唯一的，对应k8s的命名空间",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resource_quota": {
                    "$ref": "#/definitions/model.ResourceQuota"
                },
                "role_id": {
                    "description": "1表示管理员，2表示开发者",
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "model.UserRegisterReq": {
            "type": "object",
            "required": [
                "name",
                "role_id"
            ],
            "properties": {
                "email": {
                    "description": "邮箱号",
                    "type": "string"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "resource_quota": {
                    "description": "资源限制相关的,不传表示不限制",
                    "$ref": "#/definitions/model.ResourceQuota"
                },
                "role_id": {
                    "description": "角色 1表示管理员 2表示开发者",
                    "type": "integer"
                }
            }
        },
...
    }
}